Struct slog_stdlog::StdLog

pub struct StdLog;

Drain logging Records into log crate

Any Record passing through this Drain will be forwarded to log crate, just like it was created with log crate macros in the first place. The message and key-value pairs will be formated to be one string.

Caution needs to be taken to prevent circular loop where Logger installed via slog-stdlog::set_logger would log things to a StdLog drain, which would again log things to the global Logger and so on leading to an infinite recursion.