Struct slog_stdlog::StdLog

pub struct StdLog;

Drain logging Records into log crate

Using StdLog is effectively the same as using log::info!(...) and other standard logging statements.

Caution needs to be taken to prevent circular loop where Logger installed via slog-stdlog::set_logger would log things to a StdLog drain, which would again log things to the global Logger and so on leading to an infinite recursion.